simple exhange

+++ travel_bot.py
@@ -7,8 +7,8 @@ from telegram.ext import Updater, MessageHandler, Filters
 TOKEN = os.getenv('TOKEN')
 PORT = int(os.environ.get('PORT', '8443'))
 APP_NAME = os.getenv('APP_NAME')
-IDRDOL = int(os.getenv('IDRDOL'))
-IDRRUB = int(os.getenv('IDRRUB'))
+IDRDOL = float(os.getenv('IDRDOL'))
+IDRRUB = float(os.getenv('IDRRUB'))
 
 logging.basicConfig(
     format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
